+++ src/script/page/AppMain.tsx
@@ -17,16 +17,12 @@
  *
  */
 
-import {FC, useCallback, useEffect, useMemo, useState} from 'react';
+import {FC, useCallback, useEffect, useRef, useState} from 'react';
 
 import {ConnectionStatus} from '@wireapp/api-client/lib/connection/';
-import {amplify} from 'amplify';
 import {container} from 'tsyringe';
 
-import {CONV_TYPE} from '@wireapp/avs';
-import {Runtime} from '@wireapp/commons';
 import {StyledApp, THEME_ID, useMatchMedia} from '@wireapp/react-ui-kit';
-import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {CallingContainer} from 'Components/calling/CallingOverlayContainer';
 import {GroupCreationModal} from 'Components/Modals/GroupCreation/GroupCreationModal';
@@ -34,8 +30,7 @@ import {LegalHoldModal} from 'Components/Modals/LegalHoldModal/LegalHoldModal';
 import {useLegalHoldModalState} from 'Components/Modals/LegalHoldModal/LegalHoldModal.state';
 import {PrimaryModal} from 'Components/Modals/PrimaryModal';
 import {PrimaryModalComponent} from 'Components/Modals/PrimaryModal/PrimaryModal';
-import {showUserModal, UserModal} from 'Components/Modals/UserModal';
-import {iterateItem} from 'Util/ArrayUtil';
+import {UserModal} from 'Components/Modals/UserModal';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {isEscapeKey} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
@@ -48,8 +43,12 @@ import {MainContent} from './MainContent';
 import {PanelEntity, PanelState, RightSidebar} from './RightSidebar';
 import {useAppMainState, ViewType} from './state';
 import {ContentState, ListState, useAppState} from './useAppState';
+import {useInitializeApp} from './useInitializeApp';
+import {useInitializeRouter} from './useInitializeRouter';
+import {useInitSubscriptions} from './useInitSubscriptions';
 import {usePopNotification} from './usePopNotification';
 import {useWindowTitle} from './useWindowTitle';
+import {onContextMenu, shiftContent} from './utils';
 
 import {Config} from '../Config';
 import {ConversationState} from '../conversation/ConversationState';
@@ -58,13 +57,7 @@ import {Message} from '../entity/message/Message';
 import {User} from '../entity/User';
 import {ConversationError} from '../error/ConversationError';
 import {App} from '../main/app';
-import {Router} from '../router/Router';
-import {initializeRouter} from '../router/routerBindings';
 import {TeamState} from '../team/TeamState';
-import {showContextMenu} from '../ui/ContextMenu';
-import {showLabelContextMenu} from '../ui/LabelContextMenu';
-import {Shortcut} from '../ui/Shortcut';
-import {ShortcutType} from '../ui/ShortcutType';
 import {showInitialModal} from '../user/AvailabilityModal';
 import {UserState} from '../user/UserState';
 import {MainViewModel} from '../view_model/MainViewModel';
@@ -88,8 +81,6 @@ interface AppMainProps {
   mainView: MainViewModel;
 }
 
-const sidebarId = 'left-column';
-
 const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
   const apiContext = app.getAPIContext();
 
@@ -97,28 +88,28 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     throw new Error('API Context has not been set');
   }
 
+  const leftSidebarRef = useRef<HTMLDivElement | null>(null);
+
   // It's not used in ViewModel, what do to with it?
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [lastUpdate, setLastUpdate] = useState<number>();
+  const [initialMessage, setInitialMessage] = useState<Message>();
 
-  const {
-    contentState,
-    setContentState,
-    previousContentState,
-    setPreviousContentState,
-    initialMessage,
-    setInitialMessage,
-    listState,
-    setListState,
-  } = useAppState();
+  const {contentState, setContentState, previousContentState, setPreviousContentState, listState, setListState} =
+    useAppState();
 
   const {history, entity: currentEntity, close: closeRightSidebar, goTo} = useAppMainState(state => state.rightSidebar);
   const {showRequestModal} = useLegalHoldModalState();
+  const {currentView} = useAppMainState(state => state.responsiveView);
+  const isLeftSidebarVisible = currentView == ViewType.LEFT_SIDEBAR;
 
   const currentState = history.at(-1);
 
+  // To be changed when design chooses a breakpoint, the conditional can be integrated to the ui-kit directly
+  const smBreakpoint = useMatchMedia('max-width: 640px');
+
   const {repository: repositories} = app;
-  const {conversation: conversationRepository, calling: callingRepository, team: teamRepository} = repositories;
+  const {conversation: conversationRepository, team: teamRepository} = repositories;
 
   const conversationState = container.resolve(ConversationState);
   const teamState = container.resolve(TeamState);
@@ -128,15 +119,10 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     'accent_id',
     'availability',
   ]);
-  const {
-    activeConversation,
-    conversations_archived: archivedConversations,
-    conversations_unarchived: unarchivedConversations,
-  } = useKoSubscribableChildren(conversationState, [
-    'activeConversation',
-    'conversations_archived',
-    'conversations_unarchived',
-  ]);
+  const {activeConversation, conversations_archived: archivedConversations} = useKoSubscribableChildren(
+    conversationState,
+    ['activeConversation', 'conversations_archived'],
+  );
   const {isTeam: isProAccount, supportsLegalHold} = useKoSubscribableChildren(teamState, [
     'isTeam',
     'supportsLegalHold',
@@ -146,30 +132,6 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     'isActivatedAccount',
     'isTemporaryGuest',
   ]);
-
-  const visibleListItems = useMemo(() => {
-    const isStatePreferences = listState === ListState.PREFERENCES;
-
-    if (isStatePreferences) {
-      const preferenceItems = [
-        ContentState.PREFERENCES_ACCOUNT,
-        ContentState.PREFERENCES_DEVICES,
-        ContentState.PREFERENCES_OPTIONS,
-        ContentState.PREFERENCES_AV,
-      ];
-
-      if (!Runtime.isDesktopApp()) {
-        preferenceItems.push(ContentState.PREFERENCES_ABOUT);
-      }
-
-      return preferenceItems;
-    }
-
-    const hasConnectRequests = !!connectRequests.length;
-    const states: (string | Conversation)[] = hasConnectRequests ? [ContentState.CONNECTION_REQUESTS] : [];
-
-    return states.concat(unarchivedConversations);
-  }, [connectRequests.length, listState, unarchivedConversations]);
 
   useWindowTitle();
   usePopNotification(contentState, repositories.preferenceNotification);
@@ -186,12 +148,6 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
 
     closeRightSidebar();
   };
-
-  // To be changed when design chooses a breakpoint, the conditional can be integrated to the ui-kit directly
-  const smBreakpoint = useMatchMedia('max-width: 640px');
-
-  const {currentView} = useAppMainState(state => state.responsiveView);
-  const isLeftSidebarVisible = currentView == ViewType.LEFT_SIDEBAR;
 
   const updateList = (newListState: ListState, respectLastState: boolean) => {
     switch (newListState) {
@@ -212,10 +168,8 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     document.addEventListener('keydown', onKeyDownListView);
   };
 
-  const switchList = (newListState: ListState, respectLastState = true): void => {
-    const isStateChange = listState !== newListState;
-
-    if (isStateChange) {
+  const switchList = (newListState: ListState, respectLastState = true) => {
+    if (listState !== newListState) {
       hideList();
       updateList(newListState, respectLastState);
       showList(newListState);
@@ -229,34 +183,17 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     }
   };
 
-  const hideList = (): void => {
-    document.removeEventListener('keydown', onKeyDownListView);
-  };
+  const hideList = () => document.removeEventListener('keydown', onKeyDownListView);
 
   const releaseContent = (newContentState: ContentState) => {
-    const isStateConversation = previousContentState === ContentState.CONVERSATION;
     setPreviousContentState(contentState);
 
-    if (isStateConversation) {
-      const isCollectionState = ContentState.COLLECTION === newContentState;
-
-      if (!isCollectionState) {
+    if (previousContentState === ContentState.CONVERSATION) {
+      if (newContentState !== ContentState.COLLECTION) {
         conversationState.activeConversation(null);
       }
 
-      return activeConversation?.release();
-    }
-  };
-
-  const shiftContent = (hideSidebar: boolean = false) => {
-    const sidebar = document.querySelector(`#${sidebarId}`) as HTMLElement | null;
-
-    if (hideSidebar) {
-      if (sidebar) {
-        sidebar.style.visibility = 'hidden';
-      }
-    } else if (sidebar) {
-      sidebar.style.visibility = '';
+      activeConversation?.release();
     }
   };
 
@@ -266,27 +203,18 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     const isHistoryExport = newContentState === ContentState.HISTORY_EXPORT;
     const isHistoryImport = newContentState === ContentState.HISTORY_IMPORT;
 
-    return shiftContent(isHistoryExport || isHistoryImport);
+    return shiftContent(leftSidebarRef.current, isHistoryExport || isHistoryImport);
   };
 
   const checkContentAvailability = (newState: ContentState): ContentState => {
     const isStateRequests = newState === ContentState.CONNECTION_REQUESTS;
+    const hasConnectRequests = !!connectRequests.length;
 
-    if (isStateRequests) {
-      const hasConnectRequests = !!connectRequests.length;
-
-      if (!hasConnectRequests) {
-        return ContentState.WATERMARK;
-      }
-    }
-
-    return newState;
+    return isStateRequests && !hasConnectRequests ? ContentState.WATERMARK : newState;
   };
 
   const switchContent = (newContentState: ContentState): void => {
-    const isStateChange = newContentState !== contentState;
-
-    if (isStateChange) {
+    if (newContentState !== contentState) {
       releaseContent(newContentState);
       showContent(checkContentAvailability(newContentState));
     }
@@ -391,9 +319,8 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
         }
       } catch (error) {
         const conversationError = error as ConversationError;
-        const isConversationNotFound = conversationError.type === ConversationError.TYPE.CONVERSATION_NOT_FOUND;
 
-        if (isConversationNotFound) {
+        if (conversationError.type === ConversationError.TYPE.CONVERSATION_NOT_FOUND) {
           const acknowledgeModalOptions = {
             text: {
               message: t('conversationNotFoundMessage'),
@@ -411,118 +338,30 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
   );
 
   const switchPreviousContent = () => {
-    const isStateChange = previousContentState !== contentState;
-
-    if (isStateChange) {
-      const isStateRequests = previousContentState === ContentState.CONNECTION_REQUESTS;
-      if (isStateRequests) {
+    if (previousContentState !== contentState) {
+      if (previousContentState === ContentState.CONNECTION_REQUESTS) {
         switchContent(ContentState.CONNECTION_REQUESTS);
       }
 
-      const repoHasConversation = conversationState
-        .conversations()
-        .some(conversation => activeConversation && matchQualifiedIds(conversation, activeConversation));
+      const conversations = conversationState.conversations();
+      const repoHasConversation = conversations.some(
+        conversation => activeConversation && matchQualifiedIds(conversation, activeConversation),
+      );
 
       if (activeConversation && repoHasConversation && !activeConversation.is_archived()) {
-        return showConversation(activeConversation);
+        showConversation(activeConversation);
+
+        return;
       }
 
-      return switchContent(ContentState.WATERMARK);
+      switchContent(ContentState.WATERMARK);
     }
   };
 
-  const iterateActiveConversation = (reverse: boolean) => {
-    const {contentState} = useAppState.getState();
-
-    const isStateRequests = contentState === ContentState.CONNECTION_REQUESTS;
-    const activeConversationItem = isStateRequests ? ContentState.CONNECTION_REQUESTS : activeConversation;
-    const nextItem = iterateItem(visibleListItems, activeConversationItem, reverse);
-    const isConnectionRequestItem = nextItem === ContentState.CONNECTION_REQUESTS;
-
-    if (isConnectionRequestItem) {
-      return switchContent(ContentState.CONNECTION_REQUESTS);
-    }
-
-    if (nextItem) {
-      showConversation(nextItem);
-    }
-  };
-
-  const iterateActivePreference = (reverse: boolean) => {
-    let activePreference = contentState;
-    const isDeviceDetails = activePreference === ContentState.PREFERENCES_DEVICE_DETAILS;
-
-    if (isDeviceDetails) {
-      activePreference = ContentState.PREFERENCES_DEVICES;
-    }
-
-    const nextPreference = iterateItem(visibleListItems, activePreference, reverse) as ContentState;
-
-    if (nextPreference) {
-      switchContent(nextPreference);
-    }
-  };
-
-  const iterateActiveItem = (reverse = false) => {
-    const isStatePreferences = listState === ListState.PREFERENCES;
-
-    return isStatePreferences ? iterateActivePreference(reverse) : iterateActiveConversation(reverse);
-  };
-
-  const goToNext = () => iterateActiveItem(true);
-
-  const goToPrevious = () => iterateActiveItem(false);
-
-  const clickToArchive = (conversationEntity = activeConversation): void => {
+  const onArchive = (conversationEntity = activeConversation): void => {
     if (isActivatedAccount && conversationEntity) {
       mainView.actions.archiveConversation(conversationEntity);
     }
-  };
-
-  const clickToClear = (conversationEntity = activeConversation): void => {
-    if (conversationEntity) {
-      mainView.actions.clearConversation(conversationEntity);
-    }
-  };
-
-  const clickToToggleMute = (conversationEntity = activeConversation): void => {
-    if (conversationEntity) {
-      mainView.actions.toggleMuteConversation(conversationEntity);
-    }
-  };
-
-  const changeNotificationSetting = () => {
-    if (isProAccount) {
-      goTo(PanelState.NOTIFICATIONS, {entity: activeConversation});
-    } else {
-      clickToToggleMute();
-    }
-  };
-
-  const answerCall = (conversationEntity: Conversation): void => {
-    const call = callingRepository.findCall(conversationEntity.qualifiedId);
-
-    if (!call) {
-      return;
-    }
-
-    if (call.conversationType === CONV_TYPE.CONFERENCE && !callingRepository.supportsConferenceCalling) {
-      const acknowledgeModalOptions = {
-        text: {
-          message: `${t('modalConferenceCallNotSupportedMessage')} ${t('modalConferenceCallNotSupportedJoinMessage')}`,
-          title: t('modalConferenceCallNotSupportedHeadline'),
-        },
-      };
-
-      PrimaryModal.show(PrimaryModal.type.ACKNOWLEDGE, acknowledgeModalOptions);
-    } else {
-      callingRepository.answerCall(call);
-    }
-  };
-
-  const openPreferencesAbout = () => {
-    switchList(ListState.PREFERENCES);
-    switchContent(ContentState.PREFERENCES_ABOUT);
   };
 
   const openPreferencesAccount = async () => {
@@ -531,235 +370,34 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     switchContent(ContentState.PREFERENCES_ACCOUNT);
   };
 
-  const openPreferencesAudioVideo = () => {
-    switchList(ListState.PREFERENCES);
-    switchContent(ContentState.PREFERENCES_AV);
-  };
+  const openContextMenu = onContextMenu({
+    archivedConversations,
+    isProAccount,
+    conversationRepository,
+    actionsView: mainView.actions,
+    showConversation,
+    switchList,
+    listState,
+    onArchive,
+    conversationState,
+  });
 
-  const openPreferencesDevices = () => {
-    switchList(ListState.PREFERENCES);
-    switchContent(ContentState.PREFERENCES_DEVICES);
-  };
+  useInitializeApp(repositories.notification, repositories.properties, mainView.multitasking);
+  useInitializeRouter(apiContext, showConversation, switchContent, switchList, openPreferencesAccount);
+  useInitSubscriptions(
+    mainView.actions,
+    connectRequests,
+    conversationState,
+    isProAccount,
+    onArchive,
+    showConversation,
+    switchContent,
+  );
 
-  const openPreferencesOptions = () => {
-    switchList(ListState.PREFERENCES);
-    switchContent(ContentState.PREFERENCES_OPTIONS);
-  };
-
-  const openStartUI = () => {
-    switchList(ListState.START_UI);
-  };
-
-  const clickToOpenNotificationSettings = (conversationEntity = activeConversation) => {
-    if (conversationEntity) {
-      showConversation(conversationEntity, {openNotificationSettings: true});
-    }
-  };
-
-  const clickToUnarchive = (conversationEntity: Conversation): void => {
-    conversationRepository.unarchiveConversation(conversationEntity, true, 'manual un-archive').then(() => {
-      if (!archivedConversations.length) {
-        switchList(ListState.CONVERSATIONS);
-      }
-    });
-  };
-
-  const shouldHideConversation = (conversationEntity: Conversation): boolean => {
-    const isStateConversations = listState === ListState.CONVERSATIONS;
-    const isActiveConversation = conversationState.isActiveConversation(conversationEntity);
-
-    return isStateConversations && isActiveConversation;
-  };
-
-  const clickToCancelRequest = (conversationEntity: Conversation): void => {
-    const userEntity = conversationEntity.firstUserEntity();
-    const hideConversation = shouldHideConversation(conversationEntity);
-    const nextConversationEntity = conversationRepository.getNextConversation(conversationEntity);
-
-    mainView.actions.cancelConnectionRequest(userEntity, hideConversation, nextConversationEntity);
-  };
-
-  const clickToBlock = async (conversationEntity: Conversation): Promise<void> => {
-    const userEntity = conversationEntity.firstUserEntity();
-    const hideConversation = shouldHideConversation(conversationEntity);
-    const nextConversationEntity = conversationRepository.getNextConversation(conversationEntity);
-
-    await mainView.actions.blockUser(userEntity, hideConversation, nextConversationEntity);
-  };
-
-  const onContextMenu = (
-    conversationEntity: Conversation,
-    event: MouseEvent | React.MouseEvent<Element, MouseEvent>,
-  ) => {
-    const entries = [];
-
-    if (conversationEntity.isMutable()) {
-      const notificationsShortcut = Shortcut.getShortcutTooltip(ShortcutType.NOTIFICATIONS);
-
-      if (isProAccount) {
-        entries.push({
-          click: () => clickToOpenNotificationSettings(conversationEntity),
-          label: t('conversationsPopoverNotificationSettings'),
-          title: t('tooltipConversationsNotifications', notificationsShortcut),
-        });
-      } else {
-        const label = conversationEntity.showNotificationsNothing()
-          ? t('conversationsPopoverNotify')
-          : t('conversationsPopoverSilence');
-        const title = conversationEntity.showNotificationsNothing()
-          ? t('tooltipConversationsNotify', notificationsShortcut)
-          : t('tooltipConversationsSilence', notificationsShortcut);
-
-        entries.push({
-          click: () => clickToToggleMute(conversationEntity),
-          label,
-          title,
-        });
-      }
-    }
-
-    if (!conversationEntity.is_archived()) {
-      const {conversationLabelRepository} = conversationRepository;
-
-      if (!conversationLabelRepository.isFavorite(conversationEntity)) {
-        entries.push({
-          click: () => {
-            conversationLabelRepository.addConversationToFavorites(conversationEntity);
-          },
-          label: t('conversationPopoverFavorite'),
-        });
-      } else {
-        entries.push({
-          click: () => conversationLabelRepository.removeConversationFromFavorites(conversationEntity),
-          label: t('conversationPopoverUnfavorite'),
-        });
-      }
-
-      const customLabel = conversationLabelRepository.getConversationCustomLabel(conversationEntity);
-
-      if (customLabel) {
-        entries.push({
-          click: () => conversationLabelRepository.removeConversationFromLabel(customLabel, conversationEntity),
-          label: t('conversationsPopoverRemoveFrom', customLabel.name),
-        });
-      }
-
-      entries.push({
-        click: () => showLabelContextMenu(event, conversationEntity, conversationLabelRepository),
-        label: t('conversationsPopoverMoveTo'),
-      });
-    }
-
-    if (conversationEntity.is_archived()) {
-      entries.push({
-        click: () => clickToUnarchive(conversationEntity),
-        label: t('conversationsPopoverUnarchive'),
-      });
-    } else {
-      const shortcut = Shortcut.getShortcutTooltip(ShortcutType.ARCHIVE);
-
-      entries.push({
-        click: () => clickToArchive(conversationEntity),
-        label: t('conversationsPopoverArchive'),
-        title: t('tooltipConversationsArchive', shortcut),
-      });
-    }
-
-    if (conversationEntity.isRequest()) {
-      entries.push({
-        click: () => clickToCancelRequest(conversationEntity),
-        label: t('conversationsPopoverCancel'),
-      });
-    }
-
-    if (conversationEntity.isClearable()) {
-      entries.push({
-        click: () => clickToClear(conversationEntity),
-        label: t('conversationsPopoverClear'),
-      });
-    }
-
-    if (!conversationEntity.isGroup()) {
-      const userEntity = conversationEntity.firstUserEntity();
-      const canBlock = userEntity && (userEntity.isConnected() || userEntity.isRequest());
-
-      if (canBlock) {
-        entries.push({
-          click: () => clickToBlock(conversationEntity),
-          label: t('conversationsPopoverBlock'),
-        });
-      }
-    }
-
-    if (conversationEntity.isLeavable()) {
-      entries.push({
-        click: () => mainView.actions.leaveConversation(conversationEntity),
-        label: t('conversationsPopoverLeave'),
-      });
-    }
-
-    showContextMenu(event, entries, 'conversation-list-options-menu');
-  };
-
-  const initializeAppRouter = () => {
-    const router = new Router({
-      '/conversation/:conversationId(/:domain)': (conversationId: string, domain: string = apiContext.domain ?? '') => {
-        showConversation(conversationId, {}, domain);
-      },
-      '/preferences/about': () => openPreferencesAbout(),
-      '/preferences/account': () => openPreferencesAccount(),
-      '/preferences/av': () => openPreferencesAudioVideo(),
-      '/preferences/devices': () => openPreferencesDevices(),
-      '/preferences/options': () => openPreferencesOptions(),
-      '/user/:userId(/:domain)': (userId: string, domain: string = apiContext.domain ?? '') => {
-        showUserModal({domain, id: userId}, () => router.navigate('/'));
-      },
-    });
-
-    initializeRouter(router);
-    container.registerInstance(Router, router);
-  };
-
-  const initializeApp = () => {
-    repositories.notification.setContentViewModelStates(contentState, mainView.multitasking);
-
-    const redirect = localStorage.getItem(App.LOCAL_STORAGE_LOGIN_REDIRECT_KEY);
-
-    if (redirect) {
-      localStorage.removeItem(App.LOCAL_STORAGE_LOGIN_REDIRECT_KEY);
-      window.location.replace(redirect);
-    }
-
-    const conversationRedirect = localStorage.getItem(App.LOCAL_STORAGE_LOGIN_CONVERSATION_KEY);
-
-    if (conversationRedirect) {
-      const {conversation, domain} = JSON.parse(conversationRedirect)?.data;
-      localStorage.removeItem(App.LOCAL_STORAGE_LOGIN_CONVERSATION_KEY);
-      window.location.replace(`#/conversation/${conversation}${domain ? `/${domain}` : ''}`);
-    }
-
-    repositories.properties.checkPrivacyPermission().then(() => {
-      setTimeout(() => repositories.notification.checkPermission(), App.CONFIG.NOTIFICATION_CHECK);
-    });
-  };
-
-  // Initialize this only once.
   useEffect(() => {
-    initializeAppRouter();
-    initializeApp();
-
     PrimaryModal.init();
     showInitialModal(userAvailability);
 
-    amplify.subscribe(WebAppEvents.CONVERSATION.SHOW, showConversation);
-
-    amplify.subscribe(WebAppEvents.PREFERENCES.MANAGE_ACCOUNT, openPreferencesAccount);
-    amplify.subscribe(WebAppEvents.PREFERENCES.MANAGE_DEVICES, openPreferencesDevices);
-    amplify.subscribe(WebAppEvents.PREFERENCES.SHOW_AV, openPreferencesAudioVideo);
-    amplify.subscribe(WebAppEvents.SEARCH.SHOW, openStartUI);
-  }, []);
-
-  useEffect(() => {
     const conversationEntity = repositories.conversation.getMostRecentConversation();
 
     if (isTemporaryGuest) {
@@ -773,9 +411,7 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
   }, []);
 
   useEffect(() => {
-    const isStateRequests = contentState === ContentState.CONNECTION_REQUESTS;
-
-    if (isStateRequests && !connectRequests.length) {
+    if (contentState === ContentState.CONNECTION_REQUESTS && !connectRequests.length) {
       showConversation(conversationRepository.getMostRecentConversation());
     }
   }, [connectRequests, contentState, conversationRepository]);
@@ -792,16 +428,6 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
     }
   }, [supportsLegalHold]);
 
-  // Shortcuts
-  useEffect(() => {
-    amplify.subscribe(WebAppEvents.SHORTCUT.NEXT, goToNext);
-    amplify.subscribe(WebAppEvents.SHORTCUT.PREV, goToPrevious);
-    amplify.subscribe(WebAppEvents.SHORTCUT.ARCHIVE, clickToArchive);
-    amplify.subscribe(WebAppEvents.SHORTCUT.DELETE, clickToClear);
-    amplify.subscribe(WebAppEvents.SHORTCUT.NOTIFICATIONS, changeNotificationSetting);
-    amplify.subscribe(WebAppEvents.SHORTCUT.SILENCE, changeNotificationSetting); // todo: deprecated - remove when user base of wrappers version >= 3.4 is large enough
-  }, []);
-
   return (
     <StyledApp
       themeId={THEME_ID.DEFAULT}
@@ -814,7 +440,7 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
       <div id="app" className="app">
         {(!smBreakpoint || isLeftSidebarVisible) && (
           <LeftSidebar
-            answerCall={answerCall}
+            ref={leftSidebarRef}
             actionsView={mainView.actions}
             callView={mainView.calling}
             selfUser={selfUser}
@@ -822,7 +448,7 @@ const AppMain: FC<AppMainProps> = ({app, mainView, selfUser}) => {
             showConversation={showConversation}
             switchContent={switchContent}
             switchList={switchList}
-            openContextMenu={onContextMenu}
+            openContextMenu={openContextMenu}
             openPreferencesAccount={openPreferencesAccount}
             repositories={repositories}
             isFederated={mainView.isFederated}
